requirements.txt
aiogram>=3.0
python-dotenv>=1.0
aiosqlite>=0.20
Pillow>=10.0
boto3>=1.34.0
